% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Embedding.R
\name{addCustomEmbedding}
\alias{addCustomEmbedding}
\title{Add a custom embedding to an ArchRProject}
\usage{
addCustomEmbedding(
  ArchRProj = NULL,
  name = NULL,
  dfEmbedding = NULL,
  model = NULL,
  modelName = NULL,
  reducedDims = "IterativeLSI"
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{The name for the embedding to store in the given \code{ArchRProject} object.}

\item{dfEmbedding}{A data.frame, which should be alist of two variables: dimension 1 and dimension 2 for all cells.}

\item{model}{An optional calculated model. Will be saved as an RDS in the ArchRProject folder, and a link to it saved}

\item{modelName}{Name of the RDS file to be saved within the ArchR project.}

\item{reducedDims}{Name of the reducedDims space that should align with the embedding (i.e. the cell names should match). This is used to replicate how UMAP embeddings are saved.}
}
\description{
Add a custom embedding to an ArchRProject
}
