% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixFeatures.R
\name{addFragmentMatrix}
\alias{addFragmentMatrix}
\title{Add a Fragment Count Matrix to the ArrowFiles of an ArchRProject}
\usage{
addFragmentMatrix(
  ArchRProj = NULL,
  features = NULL,
  matrixName = "FragmentMatrix",
  ceiling = 500,
  verbose = TRUE,
  threads = getArchRThreads(),
  parallelParam = NULL,
  force = TRUE,
  logFile = createLogFile("addFragmentMatrix")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{features}{Optional parameter: A GRanges object of regions to count fragments within. If not provided, function will use the peakset by default (if it's available).}

\item{ceiling}{The maximum counts per feature allowed. This is used to prevent large biases in fragment counts.}

\item{verbose}{A boolean value that determines whether standard output includes verbose sections.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{force}{A boolean value indicating whether to force the "PeakMatrix" to be overwritten if it already exist in the given \code{ArchRProject}.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function, for each sample, will independently compute fragment counts (instead of insertion counts) for each peak or open window
per cell in the provided ArchRProject and save it as a 'FragmentMatrix" (unless otherwise specificed).
}
