% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DietArchRProject.R
\name{DietArchRProject}
\alias{DietArchRProject}
\title{Remove unneeded matrices from Arrow File.}
\usage{
DietArchRProject(
  ArchRProj = NULL,
  keep = "TileMatrix",
  verbose = FALSE,
  logFile = createLogFile("DietArchRProject")
)
}
\arguments{
\item{keep}{A vector of matrices to keep. You can check on all available matrices via getAvailableMatrices(). If an error occurs during copying or subsetting, some matrices may only appear within a subset of arrow files. Dieting the ArchR project down to just the available matrices will remove all matrices that do not occur in all arrow files.}

\item{verbose}{A boolean value that determines whether standard output is printed.}

\item{logFile}{The path to a file to be used for logging ArchR output.}

\item{input}{An \code{ArchRProject} object or a character vector containing the paths to the ArrowFiles to be used.}
}
\description{
This function will iterate over the arrows within an ArchRProject, only keeping the matrices selected.
Use this function if you want a minimal ArchR Project, or to recover a corrupted ArchR Project.
Sometimes, an ArchR project will become corrupted. This can occur when a multithreading application gets interrupted,
leaving some arrows with matrices, that others do not have. Removing those problematic matrices from the ArchR Project will fix any subsetting issues.
The minimum matrix to keep is is the TileMatrix - keep must include the TileMatrix, or else DietArchRProject won't run.
}
